requirements.txt
# Core Data Science
numpy>=1.26.0
pandas>=2.1.0
scipy>=1.11.0

# Data Acquisition
yfinance>=0.2.32
pandas-datareader>=0.10.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.1.0
statsmodels>=0.14.0

# Deep Learning (Optional - can skip if issues)
# tensorflow>=2.15.0
# keras>=3.0.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
kaleido>=0.2.1

# Web App
streamlit>=1.29.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
tqdm>=4.66.0

# Testing
pytest>=7.4.0

# Additional useful packages
openpyxl>=3.1.0  # For Excel file handling